перенёс статические экраны в папку screens
